one exception handler for all exceptions exception inheritance used redundant code deleted, code refactoring done
