+++ src/main/java/pl/petlovers/Pet2Vet/exceptions/ExceptionsHandler.java
@@ -10,13 +10,36 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import pl.petlovers.Pet2Vet.exceptions.forbidden_exceptions.ForbiddenAccessException;
+import pl.petlovers.Pet2Vet.exceptions.not_found_exceptions.*;
+import pl.petlovers.Pet2Vet.exceptions.unautorized_exceptions.UnauthorizedAttemptException;
 
 @Slf4j
 @ControllerAdvice
-public class OtherExceptionsHandler {
-
+public class ExceptionsHandler {
 
   static final String ERROR = "---Error: ";
+
+  @ResponseBody
+  @ExceptionHandler(ForbiddenAccessException.class)
+  @ResponseStatus(HttpStatus.FORBIDDEN)
+  String forbiddenAccessException(ForbiddenAccessException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(NotFoundInDatabaseException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String notFoundException(NotFoundInDatabaseException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(UnauthorizedAttemptException.class)
+  @ResponseStatus(HttpStatus.FORBIDDEN)
+  String unauthorizedAttemptException(UnauthorizedAttemptException exception) {
+    return exception.getMessage();
+  }
 
   @ResponseBody
   @ExceptionHandler(NullPointerException.class)
@@ -81,6 +104,49 @@ public class OtherExceptionsHandler {
   String invalidDataAccessApiUsageException(InvalidDataAccessApiUsageException exception) {
     log.error(ERROR + exception.toString());
     return exception.toString();
+  }
+
+
+  @ResponseBody
+  @ExceptionHandler(AppUserNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String appUserNotFoundHandler(AppUserNotFoundException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(MealNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String mealNotFoundHandler(MealNotFoundException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(NoteNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String noteNotFoundHandler(NoteNotFoundException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(PetNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String petNotFoundHandler(PetNotFoundException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(PetSpecieNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String petSpecieNotFoundHandler(PetSpecieNotFoundException exception) {
+    return exception.getMessage();
+  }
+
+  @ResponseBody
+  @ExceptionHandler(VaccineNotFoundException.class)
+  @ResponseStatus(HttpStatus.NOT_FOUND)
+  String vaccineNotFoundHandler(VaccineNotFoundException exception) {
+    return exception.getMessage();
   }
 
   @ResponseBody
